Initial GoodStarter Commit
(Don't blame me if this doesn't work!)

Implemented:
Basic structure of all pages with your original CSS included.

Use the "style.css" stylesheet for restyling your page. Currently I've
just ported over your old CSS, some of it I have had to change up a
little. Currently Comments is not working correctly (CSS not pulling
through) - but this will be fixed. Your CSS has been marked with your
name before it begins. Currently there is no profile page as I could not
find this in the Drive.

Currently removed most of the Razor syntax (as this is a base commit).
Will follow up with additional Razor for pages to make it more dynamic.